bz988988: clear_inode() kernel-side bug reproducer
The kernel jprobe portion of the bz988988 clear_inode bug
reproducer.  This module widens the window in which another
process might sneak in and invalidate pages after this
task obtains a list of pages, and before this task can
actually invalidate those pages.

The module places a jprobe at the instruction immediately
following the return from radix_tree_gang_lookup_slot() in
the find_get_pages() function, and performs 10,000 nop()
in a loop before returning to its previous execution
point.  This has a similar effect to an interrupt occurring
at this particular time.

Signed-off-by: Frank Sorenson <sorenson@redhat.com>