C objects
Testing out creation of objects in C.

These objects are primarily created through the use of
ELF sections, where a section contains the object, which
can then be used with dlmopen() (undocumented, but very
useful) to create multiple objects.

Testing has been very successful and informative.

See also the 'libchipper' and 'mulch' projects, which
implement the use of these objects in a well-needed
compressed-logging application.

Signed-off-by: Frank Sorenson <sorenson@redhat.com>