rhbz1792746_xfs_ilock_hang: Script to start/start tracing xfs_ilock events
Add a script to start tracing xfs_ilock-related events, to
monitor system messages for 'blocked task' alerts, and to
stop tracing (but retain captured trace data) automatically
when the 'blocked task' messages occur.

This is to allow capturing data for analysis in Red Hat bz
1792746, relating to a hang in xfs_ilock, without requiring
immediate system panic when the hang occurs.

Signed-off-by: Frank Sorenson <sorenson@redhat.com>