drop_cached_file: add a program to drop cached pages of a file
In some situations, it is desirable to invalidate
cached pages of a file in order to force the kernel
to reread those pages from the underlying storage.

The kernel provides the vm.drop_cache sysctl for
this purpose, and a filesystem unmount will also
cause this.  However, a filesystem unmount and
remount will remove all pages for all files on the
filesystem, and the sysctl does the same with a
system-wide effect, making these unsuitable due
to their wide-ranging impact.

This program uses fadvise(FADV_DONTNEED) to instruct
the kernel to discard cached pages of targeted files,
limiting the impact of the pagecache invalidation.

The program executes a single time or repeats after
a specified delay, and can run either in the
foreground or background as a daemon.

Signed-off-by: Frank Sorenson <sorenson@redhat.com>