rhbz2000342: reproducer for stale nfs file data due to nfsd bug
Due to a bug in the Linux nfs server in some RHEL major
versions 6 and 7, when a file is modified directly on the
nfs server at the same time as the file is read from an
nfs client, the nfs server can sometimes serve the old
file contents, but give the updated file metadata.

Subsequent nfs client accesses to the file will not
detect that the file has been updated because the updated
file timestamps/changeid were already returned from the
server, albeit with the stale data.

This program reproduces the bug by writing to the local
filesystem on the nfs server and reading from the same
file over nfs simultaneously.  The program will run
either on a single system (using a mount from localhost)
or on multiple systems (the writer must execute on the
nfs server in order to write to the local filesystem).

Signed-off-by: Frank Sorenson <sorenson@redhat.com>