rhbz2033552/fake_mountd: fake mountd program to behave strangely
Reproducer for autofs bug.  Replaces the rpc.mountd
program
* registers tcp 20048 and udp 20048 with rpcbind
* only EXPORT calls receive a response
* returns an error ("remote can't support version") when
  called with program version 2
* returns fixed list of 2 paths, reversing the order over
  tcp vs udp
* has a fixed 900 hosts per path
* the response is 40 KiB, and therefore too large for udp

When autofs /net/hostname is accessed, autofs will
request the export list over udp with program version
3, then 2, then 1.  Versions 3 and 1 will be too large,
and version 2 will also have an error, so autofs will
then request the export list using tcp and program
version 3.  Because the list obtained over tcp is in the
reverse order to the list over udp, the bug in autofs is
exercised. and automount segfaults.

Signed-off-by: Frank Sorenson <sorenson@redhat.com>