allocate_larger_than_size - allocate a file larger than its file size
A short demo program to show allocation of file extents
larger than the actual file size.

This is due to the behavior of the fallocate() syscall, which will
allocate space but not change the file size when the
FALLOC_FL_KEEP_SIZE mode is used.  See the manpage for fallocate(2)
or in the code for further information.


Signed-off-by: Frank Sorenson <sorenson@redhat.com>